lock on parser

+++ pipeline/external.py
@@ -13,7 +13,7 @@ import traceback
 from lccsrv.paths import *
 from legacy.extractor import *
 from subprocess import Popen, PIPE
-from manage import getParse, setParse
+from manage import getParse, setParse, getBoxerLock, setBoxerLock
 import pexpect
 import re
 
@@ -166,27 +166,31 @@ def run_annotation(request_body_dict, input_metaphors, language, task, logger, w
 	if not getParse():
 		child = pexpect.spawn(parser_proc,timeout=10)
 		setParse(True)
-
-	child.sendline(tokenizer_output)
-	reattempts = 0
-	parser_output_inter = ""
-	while reattempts < 2:
-		index = child.expect (["\n?ccg\(.*'+\)*\)\)\.\r\n", pexpect.TIMEOUT, pexpect.EOF])
-		logger.info("reattempts: %r\n" % str(reattempts))
-		if index == 0:
-			parser_output_inter = child.after
-			reattempts = 2
-		elif reattempts == 0:
-			reattempts += 1
-			child.terminate()
-			child = pexpect.spawn(parser_proc,timeout=10)
+	while True:
+		if getBoxerLock():
+			setBoxerLock(False)
 			child.sendline(tokenizer_output)
-		else:
-			logger.info("Parser not wroking\n")
-			task.log_error("\nParser not working\n")
-			reattempts = 2
-			child.terminate()
-			setParse(False)
+			reattempts = 0
+			parser_output_inter = ""
+			while reattempts < 2:
+				index = child.expect (["\n?ccg\(.*'+\)*\)\)\.\r\n", pexpect.TIMEOUT, pexpect.EOF])
+				logger.info("reattempts: %r\n" % str(reattempts))
+				if index == 0:
+					parser_output_inter = child.after
+					reattempts = 2
+				elif reattempts == 0:
+					reattempts += 1
+					child.terminate()
+					child = pexpect.spawn(parser_proc,timeout=10)
+					child.sendline(tokenizer_output)
+				else:
+					logger.info("Parser not wroking\n")
+					task.log_error("\nParser not working\n")
+					reattempts = 2
+					child.terminate()
+					setParse(False)
+			setBoxerLock(True)
+			break
 	parser_output_inter = re.sub("ccg\(\d+", "ccg(1", parser_output_inter)
 	
         logger.info("Parser output:\n%r" % parser_output_inter)
